Test graphics

+++ TetrisTypes.hs
@@ -5,7 +5,7 @@ import Graphics.Gloss
 data Block = Void | Block Color
 	deriving (Eq, Show)
 type Grid = [[Block]]
-type Position = (Int,Int)
+type Position = (Float,Float)
 
 
 type GameState = (Grid, Grid, Position, Int, Double)
